requirements/test.txt
-r dev.txt

faker==13.15.1
factory-boy==3.2.0
pytest==7.1.2
pytest-django==4.5.2
codecov==2.1.11
pytest-cov==3.0.0
